Create fetchngs.sh
This script can be placed in a folder in your .bashrc $PATH (e.g. ~/scripts/). Name it fetchngs.sh and chmod +x.

To download datasets from GEO, find the PRJNA number and call the command with "fetchngs.sh -i PRJNA123456", an optional flag of -p can make the samplesheet fit for rnaseq, etc